+++ stroom-proxy/stroom-proxy-repo/src/main/java/stroom/proxy/repo/dao/SourceDao.java
@@ -1,5 +1,6 @@
 package stroom.proxy.repo.dao;
 
+import stroom.db.util.SqliteJooqHelper;
 import stroom.proxy.repo.ProxyRepoDbConnProvider;
 
 import org.jooq.Condition;